+++ tools/rcomp-7.0.1/src/rcbinstr.cpp
@@ -117,7 +117,7 @@ void ResourceDataStream::StartOfBlockWithSizePrefix(DataType aDataTypeOfSizePref
 	{
 	NumericValue* const sizeOfBlockWhenUncompressed=new NumericValue(aDataTypeOfSizePrefix);
 	assert(sizeOfBlockWhenUncompressed!=NULL);
-	iArrayOfMarks.AppendMark(iNumberOfBytesUsed, EMarkType_StartOfBlockWithSizePrefix, (unsigned int)sizeOfBlockWhenUncompressed);
+	iArrayOfMarks.AppendMark(iNumberOfBytesUsed, EMarkType_StartOfBlockWithSizePrefix, (uintptr_t)sizeOfBlockWhenUncompressed);
 	}
 
 void ResourceDataStream::EndOfBlockWithSizePrefix()
@@ -148,11 +148,11 @@ void ResourceDataStream::StartOfCompressedUnicodeRun(int aUncompressedUnicodeSiz
 			}
 		if (insertZeroLengthCompressedUnicodeRunAtStart)
 			{
-			iArrayOfMarks.InsertMark(0, 0, EMarkType_StartOfCompressedUnicodeRun, (unsigned int)BinaryBuffer::New(0, NULL)); // mark the insertion point for the initial zero-length compressed-Unicode run
+			iArrayOfMarks.InsertMark(0, 0, EMarkType_StartOfCompressedUnicodeRun, (uintptr_t)BinaryBuffer::New(0, NULL)); // mark the insertion point for the initial zero-length compressed-Unicode run
 			iArrayOfMarks.InsertMark(1, 0, EMarkType_EndOfCompressedUnicodeRun); // mark the insertion point for the subsquent run of "other stuff"
 			}
 		}
-	iArrayOfMarks.AppendMark(iNumberOfBytesUsed, EMarkType_StartOfCompressedUnicodeRun, (unsigned int)BinaryBuffer::New(aUncompressedUnicodeSizeInBytes, aUncompressedUnicodeBuffer));
+	iArrayOfMarks.AppendMark(iNumberOfBytesUsed, EMarkType_StartOfCompressedUnicodeRun, (uintptr_t)BinaryBuffer::New(aUncompressedUnicodeSizeInBytes, aUncompressedUnicodeBuffer));
 	iContainsCompressedUnicode=true;
 	}
 
@@ -175,7 +175,7 @@ void ResourceDataStream::TwoByteAlignmentPoint()
 
 void ResourceDataStream::EnquireStreamPositionWhenKnown(unsigned long& aStreamPosition)
 	{
-	iArrayOfMarks.AppendMark(iNumberOfBytesUsed, EMarkType_EnquireStreamPositionWhenKnown, (unsigned int)&aStreamPosition);
+	iArrayOfMarks.AppendMark(iNumberOfBytesUsed, EMarkType_EnquireStreamPositionWhenKnown, (uintptr_t)&aStreamPosition);
 	}
 
 void ResourceDataStream::StreamIn(const unsigned char* aBuffer, int aNumberOfBytes)
@@ -225,13 +225,13 @@ startOfFirstPass:
 	// do a second pass to see if there are any compressed-Unicode runs that don't actually make the resource smaller (taking into consideration the preceding run-length byte(s) and any trailing run-length byte(s)) - if so, replace the EMarkType_StartOfCompressedUnicodeRun/EMarkType_EndOfCompressedUnicodeRun mark-pair with a single EMarkType_TwoByteAlignmentPoint mark and go right back to the start of the first pass
 	// do a third and final pass to actually write out the resource data to aStream
 	const int numberOfMarks=iArrayOfMarks.Size(); // caching this is done *after* the "startOfFirstPass" as ConvertCompressedRunToUncompressed (which is called just before "goto startOfFirstPass") changes the number of items in iArrayOfMarks
-	const unsigned int* lengthOfLastRun=NULL;
+	const uintptr_t* lengthOfLastRun=NULL;
 	for (int pass=0; pass<3; ++pass)
 		{
 		const int numberOfBytesBeforeFirstMark=NumberOfBytesToNextMark(-1);
 		int runLength=numberOfBytesBeforeFirstMark;
 		int uncompressedSize=numberOfBytesBeforeFirstMark;
-		unsigned int* addressToWriteSizeOfCompressedIntegerTo=NULL; // used on the first pass
+		uintptr_t* addressToWriteSizeOfCompressedIntegerTo=NULL; // used on the first pass
 		SizePrefixStack sizePrefixStack; // used on the first pass
 		int bufferIndex=0; // used on the third pass
 		if (pass==2)
@@ -261,7 +261,7 @@ startOfFirstPass:
 					{
 					aStream << *sizeOfBlockWhenUncompressed;
 					delete sizeOfBlockWhenUncompressed;
-					mark.iOtherData=(unsigned int)(NumericValue*)NULL;
+					mark.iOtherData=(uintptr_t)(NumericValue*)NULL;
 					}
 				}
 				break;
@@ -303,7 +303,7 @@ startOfFirstPass:
 					}
 				else if (pass==1)
 					{
-					const unsigned int* const lengthOfOtherStuffRun=&nextMark.iOtherData;
+					const uintptr_t* const lengthOfOtherStuffRun=&nextMark.iOtherData;
 					const bool isTheFirstCompressedUnicodeRun=!encounteredCompressedUnicode[pass];
 					const bool isTheLastCompressedUnicodeRun=(lengthOfLastRun==lengthOfOtherStuffRun);
 					if (isTheLastCompressedUnicodeRun || !isTheFirstCompressedUnicodeRun) // if this is the first compressed-Unicode run and there are others, then we can't get rid of it, hence this check
@@ -328,12 +328,12 @@ startOfFirstPass:
 					{
 					assert(pass==2);
 					runWhenUncompressed.Destroy();
-					mark.iOtherData=(unsigned int)(BinaryBuffer*)NULL;
+					mark.iOtherData=(uintptr_t)(BinaryBuffer*)NULL;
 					if (!dump)
 						aStream.WriteCompressedInteger(numberOfBytesWhenCompressed);
 					aStream.Write(iBuffer+bufferIndex, numberOfBytesWhenCompressed);
 					bufferIndex+=numberOfBytesWhenCompressed;
-					const unsigned int* const lengthOfOtherStuffRun=&nextMark.iOtherData;
+					const uintptr_t* const lengthOfOtherStuffRun=&nextMark.iOtherData;
 					if ((lengthOfLastRun!=lengthOfOtherStuffRun) || (*lengthOfOtherStuffRun>0))
 						{
 						if (!dump)
@@ -446,7 +446,7 @@ void ResourceDataStream::ConvertCompressedRunToUncompressed(int aMarkIndexOfStar
 	Mark* markForStartOfCompressedUnicodeRun=&iArrayOfMarks.MarkAt(aMarkIndexOfStartOfCompressedUnicodeRun);
 	assert(markForStartOfCompressedUnicodeRun->iMarkType==EMarkType_StartOfCompressedUnicodeRun);
 	BinaryBuffer& runWhenUncompressed=*(BinaryBuffer*)markForStartOfCompressedUnicodeRun->iOtherData;
-	markForStartOfCompressedUnicodeRun->iOtherData=(unsigned int)(BinaryBuffer*)NULL;
+	markForStartOfCompressedUnicodeRun->iOtherData=(uintptr_t)(BinaryBuffer*)NULL;
 	Mark* markForEndOfCompressedUnicodeRun=&iArrayOfMarks.MarkAt(aMarkIndexOfStartOfCompressedUnicodeRun+1);
 	assert(markForEndOfCompressedUnicodeRun->iMarkType==EMarkType_EndOfCompressedUnicodeRun);
 	const int bufferPositionOfStartOfCompressedUnicodeRun=markForStartOfCompressedUnicodeRun->iBufferPosition;
